The loader accepts additional fields calculated from the base data collection. For this, the "task_details.json" contains elements like below:
    {
        "field_name": "@AV_PREFETCH_WAIT_TIME",
        "field_valid": "true",
        "field_type": "decimal(5,2)",
        "field_formula": "CASE WHEN PREFETCH_WAITS = 0 THEN 0 ELSE ( PREFETCH_WAIT_TIME / PREFETCH_WAIT_TIME ) end"
    }

After creation of the DELTA table, fields defined this way may be appended. This way, we can get a formula on the delta values (rather than a delta value of the formula). The '@' is a convention used to indicate that this is a field added to the DELTA table.

The setup script had to be modified to accept these json elements.